PointToBrick with empty pointedObject
According to Catroid's
org.catrobat.catroid.content.actions.PointToAction class the
pointedObject gets set to the same object the Brick is in: "if
(pointedSprite == null) { pointedSprite = this.sprite; }"

This scenerio happens when you have a PointToBrick pointing to an
object which has been deleted (after the PointToBrick has been set up).

The resulting node in the XML file has no children and looks like:
"<brick type="PointToBrick"/ >"

Since Catroid is producing those XML files in certain situations (as
mentioned above), we have to consider this case in our parser and
serializer as well.