IOS-388 #done
Removed test-target from release build as it is not possible to build
it with the build option ENABLE_TESTABILITY set to „no“. This option is
necessary to enable all optimizations. Therefore testing and
optimization is not possible at the same time. Disabled compilation of
the tests in release too. Solution is recommended by Apple.